Fix scattered packages
Somehow magit caused several hunks to get shuffled around. Or was it
my fault? I don't think so, but then again I don't have a clear idea
what exactly happened. In any case, it would be a pain in the ass to
fixup the relevant commits, so I'm doing it all here.