Modify defkey to not require quotation marks
Instead of (defkey "M-s" save-buffer) and (defkey "C-x C-b" ibuffer),
put (defkey M-s save-buffer) and (defkey (C-x C-b) ibuffer). Is this a
frivolous and extravagant macro? Maybe. Is it inexcusably lazy to want
to type keybindings without quotation marks? Maybe.

It would be nice to be able to do all the bindings in one go, as
in (defkeys M-s save-buffer (C-x C-b) ibuffer).