Change config load order
The order in which config files are loaded should basically be from
least likely to need debugging to most. If there an error in keys, for
instance, and custom hasn't been loaded, then the error in keys will
need to be diagnosed and fixed against a bright white background,
which sucks.