Change god-mode modifier keys
I stopped using god-mode much because I had j assigned to meta, when I
really wanted to use j for avy. Then I realized -- m, i, and [ can't
be used with C anyway (which is really annoying), so why not use them
as god modifiers? I don't want to use w for anything, so I'll make
that a modifier too.

Rationale for choices:

m - Meta, that's easy.

i - My most frequently used Super command is s-k for kill-this-buffer,
and i k is an easy chord.

w - On my Mac I use the Fn key for Hyper. That's only on the left
side of the keyboard, so most of my Hyper bindings have the key on the
right side -- H-p, H-k, etc. w is on the left side, it works fine for
this.

[ - C-M commands seem to be most used with keys on the left side of
the keyboard -- C-M-f, C-M-a, etc. [ is on the right side, so there
you go.